moves pieces and player into their own respective packages.
