+++ src/pieces/Pawn.java
@@ -1,0 +1,124 @@
+
+package pieces;
+
+import java.awt.Image;
+import java.awt.Point;
+
+public class Pawn extends Piece {
+
+    private boolean myseen = false;
+    private boolean movedbefore = false;
+
+    /** Creates a new instance of Pawn */
+    public Pawn(String NameIcon, int startX, int startY) {
+
+        pieceIcon = new PieceIcon(NameIcon);
+
+        X = startX;
+        Y = startY;
+        p.x = X;
+        p.y = Y;
+    }
+
+    public boolean Canmove(int x, int y, String typeColor) {
+
+        if ((typeColor.equals("black"))) {
+            if ((((y - 1 == Y) && (x == (X)))) /* &&!Check_Solider_Sees(x,y) */) {
+
+                return true;
+
+            } else if ((((y - 2 == Y) && (x == (X)))) && !movedbefore) {
+
+                return true;
+            } else if ((y - 1 == Y && x + 1 == (X) || (y - 1 == Y && x - 1 == (X))) && myseen) {
+                return true;
+            }
+
+            else
+                return false;
+        }
+
+        else if (typeColor == "white") {
+            if (((y + 1 == Y) && (x == (X))) /* &&!Check_Solider_Sees(x,y) */) {
+                return true;
+            } else if ((((y + 2 == Y) && (x == (X)))) && !movedbefore) {
+                return true;
+            } else if ((y + 1 == Y && x + 1 == (X) || (y + 1 == Y && x - 1 == (X))) && myseen) {
+                return true;
+            }
+
+            else
+                return false;
+        }
+        return false;
+
+    }
+
+    public boolean PieceInMYway(int x, int y, Point othersPostion, String typeColor) {
+        if (Y - y == 2 || Y - y == -2) {
+            if ((typeColor.equals("black"))) {
+
+                if ((((y - 1 == othersPostion.y) && (x == (othersPostion.x)))) && !movedbefore) {
+                    return true;
+                } else
+                    return false;
+            }
+
+            else if (typeColor.equals("white")) {
+
+                if (((y + 1 == othersPostion.y) && (x == (othersPostion.x)) && !movedbefore)) {
+
+                    return true;
+
+                } else
+                    return false;
+            }
+        }
+
+        return false;
+    }
+
+    public void toOld(Point Old) {
+
+        p.x = Old.x;
+        p.y = Old.y;
+
+    }
+
+    public void setMYseen(boolean newBoolean) {
+        myseen = newBoolean;
+    }
+
+    public boolean returnMyseen() {
+        return myseen;
+    }
+
+    public boolean setSeenbychecking(Point newP, String Color) {
+        myseen = false;
+        if ((Color.equals("black"))) {
+            if ((newP.y - 1 == Y && newP.x + 1 == (X) || (newP.y - 1 == Y && newP.x - 1 == (X)))) {
+
+                myseen = true;
+                return true;
+            } else
+                return false;
+        } else if (Color.equals("white")) {
+            if ((newP.y + 1 == Y && newP.x + 1 == (X) || (newP.y + 1 == Y && newP.x - 1 == (X)))) {
+                myseen = true;
+
+                return true;
+            } else
+                return false;
+        }
+        return false;
+    }
+
+    public Point GeneratePossible_Moves() {
+        return new Point();
+    }
+
+    public String Tell_me() {
+        return "Soldier= (" + p.x + ',' + p.y + ")";
+    }
+
+}
